# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Primea Ticketing API",
        "description": "API documentation for Primea Ticketing System - A comprehensive event ticketing platform",
        "contact": {
            "email": "support@primea.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/webhooks/airtel": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Handle Airtel Money webhook",
                "description": "Processes payment status updates from Airtel Money gateway",
                "operationId": "handleAirtelWebhook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reference": {
                                        "description": "Payment reference from the original payment request",
                                        "type": "string",
                                        "example": "PAY_12345_67890"
                                    },
                                    "status": {
                                        "description": "Payment status from Airtel Money",
                                        "type": "string",
                                        "enum": [
                                            "success",
                                            "failed",
                                            "cancelled",
                                            "pending"
                                        ],
                                        "example": "success"
                                    },
                                    "transaction_id": {
                                        "description": "Airtel Money transaction ID",
                                        "type": "string",
                                        "example": "TXN_AM_123456789",
                                        "nullable": true
                                    },
                                    "amount": {
                                        "description": "Transaction amount",
                                        "type": "number",
                                        "format": "float",
                                        "example": 75.5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Webhook Airtel traité"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing reference",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reference missing"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/webhooks/moov": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Handle Moov Money webhook",
                "description": "Processes payment status updates from Moov Money gateway",
                "operationId": "handleMoovWebhook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reference": {
                                        "description": "Payment reference from the original payment request",
                                        "type": "string",
                                        "example": "PAY_12345_67890"
                                    },
                                    "status": {
                                        "description": "Payment status from Moov Money",
                                        "type": "string",
                                        "enum": [
                                            "success",
                                            "failed",
                                            "cancelled",
                                            "pending"
                                        ],
                                        "example": "success"
                                    },
                                    "transaction_id": {
                                        "description": "Moov Money transaction ID",
                                        "type": "string",
                                        "example": "TXN_MOOV_123456789",
                                        "nullable": true
                                    },
                                    "amount": {
                                        "description": "Transaction amount",
                                        "type": "number",
                                        "format": "float",
                                        "example": 75.5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Webhook Moov traité"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing reference",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reference missing"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/webhooks/card": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Handle Card payment webhook",
                "description": "Processes payment status updates from card payment gateway",
                "operationId": "handleCardWebhook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reference": {
                                        "description": "Payment reference from the original payment request",
                                        "type": "string",
                                        "example": "PAY_12345_67890"
                                    },
                                    "status": {
                                        "description": "Payment status from card gateway",
                                        "type": "string",
                                        "enum": [
                                            "success",
                                            "failed",
                                            "cancelled",
                                            "pending",
                                            "declined"
                                        ],
                                        "example": "success"
                                    },
                                    "transaction_id": {
                                        "description": "Card gateway transaction ID",
                                        "type": "string",
                                        "example": "TXN_CARD_123456789",
                                        "nullable": true
                                    },
                                    "amount": {
                                        "description": "Transaction amount",
                                        "type": "number",
                                        "format": "float",
                                        "example": 75.5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Webhook Carte traité"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing reference",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reference missing"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}